fix this 3.1 error: ’+’ is deprecated: Mixed-type addition is deprecated. Please use explicit type conversion.

+++ Sources/SwiftRandom/URandom.swift
@@ -33,7 +33,7 @@ public class URandom: Random {
 	private func read(numBytes: Int) -> [Int8] {
 		// Initialize an empty array with numBytes+1 for null terminated string
 		var bytes = [Int8](repeating: 0, count: numBytes + 1)
-		fgets(&bytes, numBytes + 1, file)
+		fgets(&bytes, Int32(numBytes) + Int32(1), file)
 
 		bytes.removeLast()
